Increase the inactivity timeout
There is a commonly observed issue when using the GSS through a DBus
proxy in which a method call fails due to an UnknownService error. This
happens because the GSS is shutting down while receiving the metehod
call.
Since the inactivity timeout for a service type GApplication is 10
seconds but, as observed during the quests, queries to the GSS usually
happen multiple times for periods longer than that, it means that
having a 10 seconds timeout may lead to more chances of hitting the
shutdown issue then a longer timeout.

Thus, this patch increases the inactivity timeout of the GSS, and while
this doesn't prevent the same issue described above from happpening, it
should make it less frequent.

https://phabricator.endlessm.com/T26343