Reset the inactivity timeout on every method call
There is a commonly observed issue when using the GSS through a DBus
proxy in which a method call fails due to an UnknownService error. This
happens because the GSS is shutting down while receiving the metehod
call.

In the previous commit, the inactivity timeout was increased, but this
timeout is counted from the start up of the application, thus, if a
client repeatedly uses the GSS for a duration of the inactivity timeout
there are high chances that the call-during-shutdown issue will happen.

This patch prevents the above issue by resetting the inactivity timeout
on every method call, so apps can keep using the GSS repeatedly even
beyond an initial timeout count.

Note that this doesn't prevent the call-during-shutdown issue, but
makes it harder to happen, as it means that it'll happen only if two
calls happen with exactly the inactivity timeout between them.

https://phabricator.endlessm.com/T26343