Move inactivity-timeout initial reset to startup
The hold/release logic for correctly resetting the initial
inactivity-timeout use was set in the constructor, meaning it would
hold any instance of the GSS that amount of time. This is not needed
except for the main instance so, these changes move the mentioned
logic into the startup method for that effect.

https://phabricator.endlessm.com/T26744