Add CLI with option for reloading the state
Since the GSS stays alive for a few minutes after its last use, one has
to reload or relaunch it after manually deploying a GSS file. In order
to make that easier, these changes add a CLI to the GSS with a --reload
option for that effect.

Due to the way of using the CLI, i.e. a main instance is running and the
user runs a secondary instance with the --reload option that will be
processed in the main instance, the GSS cannot be set up as a "service"
type GApplication because this makes a registration attempt fail and
we cannot easily handle the CLI options that way. So this patch also
drops that property, making the GSS run as non-service app. This should
have no impact in how the GSS actually performs.

https://phabricator.endlessm.com/T26744